Fixing validation button, bugs: post & css grid + styling
Still need to finnish the server part of the validating part.
The css grid for some reason cannot be put on the individual routes but needs to be put on the route-view itself. We need to look how to fix this without hopfully having to add more empty divs (bad for accessibilty)